pyproject.toml
[tool.poetry]
name = "finalproject-zhabinskaya-m25-555"
version = "0.1.0"
description = "ValutaTrade Hub"
authors = ["Victoria Zhabinskaya <zhabinskayavictoria@yandex.ru>"]
readme = "README.md"
packages = [
     { include = "valutatrade_hub" },
]

[tool.poetry.scripts]
project = "main:main"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
prettytable = "^3.16.0"
requests = "^2.32.3"
python-dotenv = "^1.0.0"

[tool.valutatrade]
DATA_DIR = "data"
USERS_FILE = "data/users.json"
PORTFOLIOS_FILE = "data/portfolios.json"
RATES_FILE = "data/rates.json"
EXCHANGE_RATES_FILE = "data/exchange_rates.json"
RATES_TTL_SECONDS = 300
DEFAULT_BASE_CURRENCY = "USD"
LOG_DIR = "logs"
LOG_FILE = "valutatrade.log"
ACTIONS_LOG_FILE = "actions.log" 
PARSER_LOG_FILE = "parser.log"
LOG_LEVEL = "INFO"
LOG_FORMAT = "text"
LOG_MAX_SIZE_MB = 10
LOG_BACKUP_COUNT = 3

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.poetry.group.dev.dependencies]
ruff = ">=0.14.0,<0.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
